refac(math): move the code of binary fields
