+++ tachyon/math/finite_fields/binary_fields/binary_field_traits_forward.h
@@ -1,0 +1,11 @@
+#ifndef TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_FIELD_TRAITS_FORWARD_H_
+#define TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_FIELD_TRAITS_FORWARD_H_
+
+namespace tachyon::math {
+
+template <typename T, typename SFINAE = void>
+struct BinaryFieldTraits;
+
+}  // namespace tachyon::math
+
+#endif  // TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_FIELD_TRAITS_FORWARD_H_
