+++ tachyon/math/finite_fields/binary_fields/binary_tower_operations.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a Apache-2.0 style license that
 // can be found in the LICENSE.ulvetanna file.
 
-#ifndef TACHYON_MATH_FINITE_FIELDS_BINARY_TOWER_OPERATIONS_H_
-#define TACHYON_MATH_FINITE_FIELDS_BINARY_TOWER_OPERATIONS_H_
+#ifndef TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_TOWER_OPERATIONS_H_
+#define TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_TOWER_OPERATIONS_H_
 
 #include <stddef.h>
 #include <stdint.h>
@@ -380,4 +380,4 @@ struct BinaryTowerOperations<F, std::enable_if_t<(F::kBits >= 16)>> {
 
 }  // namespace tachyon::math
 
-#endif  // TACHYON_MATH_FINITE_FIELDS_BINARY_TOWER_OPERATIONS_H_
+#endif  // TACHYON_MATH_FINITE_FIELDS_BINARY_FIELDS_BINARY_TOWER_OPERATIONS_H_
