+++ tachyon/math/finite_fields/generator/binary_field_generator/binary_field_config.h.tpl
@@ -1,4 +1,7 @@
 // clang-format off
+#include <stddef.h>
+#include <stdint.h>
+
 #include "tachyon/export.h"
 %{if NeedsBigInt}
 #include "tachyon/math/base/big_int.h"
