+++ tachyon/math/finite_fields/generator/binary_field_generator/binary_field_cpu.h.tpl
@@ -1,7 +1,7 @@
 // clang-format off
 #include "%{config_header_path}"
 
-#include "tachyon/math/finite_fields/binary_field.h"
+#include "tachyon/math/finite_fields/binary_fields/binary_field.h"
 
 namespace %{namespace} {
